automatically backed up by learn

+++ lib/bachelor.rb
@@ -43,7 +43,7 @@ counter = 0
 sum_of_ages = 0 
   data[season].each do|contestant_hash|
     counter += 1 
-    sum_of_ages += contestant_hash["age"].to_f
+    sum_of_ages += contestant_hash["age"].to_i
   average_age = sum_of_ages / counter
     return average_age.round
   end
